% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_list.R
\docType{data}
\name{book_list}
\alias{book_list}
\title{Ctext.org Book List}
\format{data.frame object.}
\usage{
book_list
available books <- unique(book_list$book)
}
\description{
An internal dataframe used for the get_text and get_books functions for downloading
books from the ctext.org API. Contains information for name of book, chapter, chapter number, and genre.
Indludes over 130 different books available on ctext.org.
}
\keyword{datasets}
