% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_books}
\alias{get_books}
\title{Get an entire book that is in book_list}
\usage{
get_books(...)
}
\arguments{
\item{a}{vector detailing a book or multiple books.}
}
\value{
Returns a dataframe with 4 columns: text, book, chapter in English, & chapter in Chinese.
In the case that the connection experiences a problem or you meet your daily API imit,
an NA value will be returned in the "text" column.
}
\description{
A wrapper around get_text() that downloads an entire book.
}
\examples{
get_books("analects")
get_books("dao-de-jing")
#Be careful not to download too many books at once - you may hit the API limit.
my_booklist <- c("analects", "mengzi", "zhuangzi")
my_books <- get_books(my_booklist)

# For "Mozi" you will only be allowed to download Chapter 1 and none of the other books,
#unless you register.

}
