% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_calls}
\alias{get_calls}
\title{Get total API call count}
\usage{
get_calls()
}
\value{
an integer
}
\description{
Get the amount of times you have called the ctext.org API.
}
\examples{
get_calls()

}
