% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_chapter}
\alias{get_chapter}
\title{get_chapter - Download a single chapter from ctext.org}
\usage{
get_chapter(book, chapter = NA)
}
\arguments{
\item{book}{name of book}

\item{chapter}{name or number of chapter in book}
}
\value{
Returns a dataframe with 4 columns: text, book, chapter name in English, & chapter name in Chinese.
  In the case for the dao-de-jing, chapter param not needed.
  If you reach the API call limit or the connection has an error, function will retry 3 times.
   If still unsuccessful, an NA value will be returned in the "text" column.
}
\description{
Unvectorized function. Used as a building block for get_chapters and get_books.
}
\examples{
get_chapter("analects", "xue-er")
get_chapter("analects", 1)
get_chapter("dao-de-jing")
}
