% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_text.R
\name{get_text}
\alias{get_text}
\title{download a chapter of a Chinese book on ctext.org}
\usage{
get_text(book, chapter = NA)
}
\arguments{
\item{book}{name of book}

\item{chapter}{name or number of chapter in book}
}
\value{
Returns a dataframe with 4 columns: text, book, chapter name in English, & chapter name in Chinese.
  In the case for the dao-de-jing, only book arg is needed. In the case you reach the API call limit or
  the connection has an error, an NA value will be returned in the "text" column.
}
\description{
Download a chapter's worth of text from a Chinese book available in book_list.
}
\examples{
get_text("analects", "xue-er")
get_text("analects", 1)
get_text("dao-de-jing")
}
